+++ src/App.css
@@ -22,6 +22,7 @@ article, aside, details, figcaption, figure, footer, header, hgroup, menu, nav, 
 
 body {
   line-height: 1;
+  background-color: #fafafa;
 }
 
 ol, ul {
