+++ src/data/Entrys.js
@@ -1,0 +1,36 @@
+const ENTRYS = [
+  {
+    id: '8392473984',
+    title: '記事タイトル1',
+    date: '2017-11-20',
+    author: 'kentatan',
+    body: '記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。',
+    image_url: 'https://placehold.jp/800x800.png'
+  },
+  {
+    id: '2381273217',
+    title: '記事タイトル2記事タイトル2',
+    date: '2017-11-21',
+    author: 'jyasukaw',
+    body: '記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。',
+    image_url: 'https://placehold.jp/800x800.png'
+  },
+  {
+    id: '49873434374',
+    title: '記事タイトル3記事タイトル3記事タイトル3',
+    date: '2017-11-22',
+    author: 'kentatan',
+    body: '記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。',
+    image_url: 'https://placehold.jp/800x800.png'
+  },
+  {
+    id: '98848234923',
+    title: '記事タイトル4記事タイトル4記事タイトル4記事タイトル4',
+    date: '2017-11-23',
+    author: 'jyasukaw',
+    body: '記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。記事の内容です。',
+    image_url: 'https://placehold.jp/800x800.png'
+  }
+]
+
+export default ENTRYS;
